refactor(main): extract a composition root from index.ts
`src/main/index.ts` was 1464 lines and the highest-churn file in the repo,
with `app.whenReady().then(async () => …)` spanning ~500 lines as an
anonymous callback that built 20+ services into 27 module-scope `let`s.

Because construction had no named order, anything wired before the service
it reads had to take a function instead of a value — the footgun
`src/main/AGENTS.md` documents, since `() => X | null` and `X | null` both
type-check at every call site. It also left 23 unreachable
"The account service is not ready." guards and a hand-maintained shutdown
sequence nothing linked to construction.

The callback moves whole into `createApplicationServices()` — one function,
not stages, so the control-flow narrowing that lets a `const` reach a
non-null parameter keeps working. Each service registers its own teardown
step beside its construction.
